refactor: move generatePriceEvents to its own file
