+++ src/game-logic/reducers/index.test.js
@@ -3,8 +3,6 @@ import {
   ACHIEVEMENT_COMPLETED,
   LOAN_INCREASED,
   LOAN_PAYOFF,
-  PRICE_CRASH,
-  PRICE_SURGE,
 } from '../../templates'
 import {
   COW_GESTATION_PERIOD_DAYS,
@@ -16,7 +14,7 @@ import {
   PURCHASEABLE_COW_PENS,
   STORAGE_EXPANSION_AMOUNT,
 } from '../../constants'
-import { huggingMachine, sampleCropItem1 } from '../../data/items'
+import { huggingMachine } from '../../data/items'
 import { sampleRecipe1 } from '../../data/recipes'
 import { genders, standardCowColors } from '../../enums'
 import {
@@ -24,7 +22,6 @@ import {
   getCostOfNextStorageExpansion,
   getCowValue,
   getPlotContentFromItemId,
-  getPriceEventForCrop,
 } from '../../utils'
 
 import * as fn from './'
@@ -44,81 +41,6 @@ jest.mock('../../constants', () => ({
   CROW_CHANCE: 0,
   PRECIPITATION_CHANCE: 0,
 }))
-
-describe('generatePriceEvents', () => {
-  describe('price event already exists', () => {
-    test('no-ops', () => {
-      jest.spyOn(Math, 'random').mockReturnValue(1)
-      const inputState = {
-        newDayNotifications: [],
-        priceCrashes: {
-          [sampleCropItem1.id]: {
-            itemId: sampleCropItem1.id,
-            daysRemaining: 1,
-          },
-        },
-        priceSurges: {},
-      }
-      const { priceCrashes, priceSurges } = fn.generatePriceEvents(inputState)
-
-      expect(priceCrashes).toEqual(inputState.priceCrashes)
-      expect(priceSurges).toEqual(inputState.priceSurges)
-    })
-  })
-
-  describe('price event does not already exist', () => {
-    let state
-
-    beforeEach(() => {
-      jest.spyOn(Math, 'random').mockReturnValue(0)
-
-      jest.resetModules()
-      jest.mock('../../data/levels', () => ({
-        levels: [
-          {
-            id: 0,
-          },
-          {
-            id: 1,
-            unlocksShopItem: 'sample-crop-seeds-1',
-          },
-        ],
-        itemUnlockLevels: {},
-      }))
-      const { generatePriceEvents } = jest.requireActual('./')
-      state = generatePriceEvents({
-        newDayNotifications: [],
-        priceCrashes: {},
-        priceSurges: {},
-        itemsSold: { 'sample-crop-1': Infinity },
-      })
-    })
-
-    test('generates a price event', () => {
-      const priceEvents = {
-        [sampleCropItem1.id]: getPriceEventForCrop(sampleCropItem1),
-      }
-
-      expect(state).toContainAnyEntries([
-        ['priceCrashes', priceEvents],
-        ['priceSurges', priceEvents],
-      ])
-    })
-
-    test('shows notification', () => {
-      expect(state.newDayNotifications).toIncludeAnyMembers([
-        {
-          message: PRICE_CRASH`${sampleCropItem1}`,
-          severity: 'warning',
-        },
-        {
-          message: PRICE_SURGE`${sampleCropItem1}`,
-          severity: 'success',
-        },
-      ])
-    })
-  })
-})
 
 describe('updatePriceEvents', () => {
   test('updates price events', () => {
